+++ source/client/client.h
@@ -411,7 +411,7 @@ void CL_GameModule_RenderView( float stereo_separation );
 void CL_GameModule_GetEntitySpatilization( int entnum, vec3_t origin, vec3_t velocity );
 void CL_GameModule_InputFrame( int frameTime );
 void CL_GameModule_ClearInputState( void );
-uint8_t CL_GameModule_GetButtonBits( void );
+unsigned CL_GameModule_GetButtonBits( void );
 void CL_GameModule_AddViewAngles( vec3_t viewAngles );
 void CL_GameModule_AddMovement( vec3_t movement );
 void CL_GameModule_MouseMove( int dx, int dy );
